+++ admin/routes/tech-support.js
@@ -4,12 +4,18 @@ const roles = require('../lib/consts/roles')
 const express = require('express')
 const router = express.Router()
 const isAuthenticated = require('../authentication/middleware')
-const techSupportHomeController = require('../controllers/tech-support-home')
+const techSupportController = require('../controllers/tech-support')
 
 router.get(
   ['/home'],
   isAuthenticated([roles.techSupport]),
-  (req, res, next) => techSupportHomeController.getTechSupportHomePage(req, res, next)
+  (req, res, next) => techSupportController.getHomePage(req, res, next)
+)
+
+router.get(
+  ['/checkview'],
+  isAuthenticated([roles.techSupport]),
+  (req, res, next) => techSupportController.getCheckViewPage(req, res, next)
 )
 
 module.exports = router
